// frontend/src/lib/api/models/recordScheduleAdminSearchResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Stream Archive API
 * 멀티 플랫폼 스트리밍 녹화 시스템 API
 * OpenAPI spec version: 1.0.0
 */
import type { ChannelInfo } from "./channelInfo";
import type { RecordScheduleAdminSearchResponsePlatformType } from "./recordScheduleAdminSearchResponsePlatformType";
import type { RecordScheduleAdminSearchResponseScheduleType } from "./recordScheduleAdminSearchResponseScheduleType";
import type { RecordScheduleAdminSearchResponseRecordQuality } from "./recordScheduleAdminSearchResponseRecordQuality";

/**
 * 녹화 스케줄 목록 응답 (관리자)
 */
export interface RecordScheduleAdminSearchResponse {
  /** 녹화 스케줄 ID */
  id: number;
  channel: ChannelInfo;
  /** 플랫폼 유형 (CHZZK/SOOP/TWITCH/YOUTUBE) */
  platformType: RecordScheduleAdminSearchResponsePlatformType;
  /** 녹화 스케줄 유형 (ONCE/ALWAYS/WEEKLY/SPECIFIC) */
  scheduleType: RecordScheduleAdminSearchResponseScheduleType;
  /** 스케줄 값 */
  value: string;
  /** 녹화 화질 */
  recordQuality: RecordScheduleAdminSearchResponseRecordQuality;
  /** 우선순위 */
  priority: number;
  /** 자동 소장 여부 */
  autoArchive: boolean;
  /** 생성 일시 */
  createdAt: string;
  /** 수정 일시 */
  updatedAt: string;
}